feat: support for basic recursion descent parsing of expressions
I'm slow, but it was quite the puzzle !
Parsers are not that easy to wrap your head around.

I figured out that my naive strategy is named recursion descent,
and it's not optimal but I guess it'll have to do for now.

Stratifying the grammar to handle precedence is verbose, but it works.

/spend 6h